+++ gpustat/__init__.py
@@ -2,7 +2,16 @@
 The gpustat module.
 """
 
-__version__ = '1.1.0.dev0'
+# isort: skip_file
+try:
+    from ._version import version as __version__
+    from ._version import version_tuple as __version_tuple__
+except (ImportError, AttributeError) as ex:
+    raise ImportError(
+        "Unable to find `gpustat.__version__` string. "
+        "Please try reinstalling gpustat; or if you are on a development "
+        "version, then run `pip install -e .` and try again."
+    ) from ex
 
 from .core import GPUStat, GPUStatCollection
 from .core import new_query