+++ CHANGELOG.md
@@ -16,6 +16,7 @@ Changelog for `gpustat`
 
 - Handle Lost GPU and Unknown Error situations (#81, #125)
 - Print a summary of the error message when an error happens.
+- Use setuptools-scm to auto-generate `__version__` string.
 - Add Python 3.11 to CI.
 
 ### Bugfix